+++ frontend/src/app/mocks/services/otherParts-mock/otherParts.handler.ts
@@ -29,9 +29,9 @@ import { supplierPartsAsPlannedAssets } from './supplierPartsAsPlanned.model';
 export const otherPartsAsBuiltHandlers = [
   rest.get(`*${environment.apiUrl}/assets/as-built`, (req, res, ctx) => {
     const pagination = extractPagination(req);
-    const owner = req.url.searchParams.get('owner');
+    const ownerSearchQuery = req.url.searchParams.get('filter').replace("owner,EQUAL,", "");
 
-    switch (owner) {
+    switch (ownerSearchQuery) {
       case 'SUPPLIER':
         return res(ctx.status(200), ctx.json(mockSupplierAssets));
 
@@ -44,7 +44,7 @@ export const otherPartsAsBuiltHandlers = [
 
 export const otherPartsAsBuiltHandlersTest = [
   rest.get(`*${environment.apiUrl}/assets/as-built`, (req, res, ctx) => {
-    const owner = req.url.searchParams.get('owner');
+    const owner = req.url.searchParams.get('filter').replace("owner,EQUAL,", "");
 
     switch (owner) {
       case 'SUPPLIER':
@@ -60,7 +60,7 @@ export const otherPartsAsBuiltHandlersTest = [
 export const otherPartsAsPlannedHandlers = [
   rest.get(`*${environment.apiUrl}/assets/as-planned`, (req, res, ctx) => {
     const pagination = extractPagination(req);
-    const owner = req.url.searchParams.get('owner');
+    const owner = req.url.searchParams.get('filter').replace("owner,EQUAL,", "");
 
     switch (owner) {
       case 'SUPPLIER':
@@ -76,7 +76,7 @@ export const otherPartsAsPlannedHandlers = [
 export const otherPartsAsPlannedHandlersTest = [
   rest.get(`*${environment.apiUrl}/assets/as-planned`, (req, res, ctx) => {
     const pagination = extractPagination(req);
-    const owner = req.url.searchParams.get('owner');
+    const owner = req.url.searchParams.get('filter').replace("owner,EQUAL,", "");
 
     switch (owner) {
       case 'SUPPLIER':